data_api/pyproject.toml
[project]
name = "data-api"
version = "0.1.0"
description = "Message Intent Analysis - Gemini API integration"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
]

[project.scripts]
test-gemini = "test_gemini_api:main"
